Fixed sequence contains no matching element error

+++ SteamCSharp/SteamCSharp.Entities/Statistics/Users/Games/CounterStrike/GlobalOffensive.cs
@@ -6,43 +6,43 @@ namespace SteamCSharp.Entities.Statistics.Users.Games.CounterStrike
 {
     public class GlobalOffensive
     {
-        public int TotalKills { get; set; }
-        public int TotalKillsHeadshot { get; set; }
-        public int TotalKillsKnifeFight { get; set; }
-        public int TotalKillsPistolRound { get; set; }
-        public int TotalKillsEnemyWeapon { get; set; }
-        public int TotalKillsEnemyBlinded { get; set; }
-        public int TotalKillsAgainstZoomedSniper { get; set; }
+        public double TotalKills { get; set; }
+        public double TotalKillsHeadshot { get; set; }
+        public double TotalKillsKnifeFight { get; set; }
+        public double TotalKillsPistolRound { get; set; }
+        public double TotalKillsEnemyWeapon { get; set; }
+        public double TotalKillsEnemyBlinded { get; set; }
+        public double TotalKillsAgainstZoomedSniper { get; set; }
         public double KillDeathRatio { get; set; }
-        public int TotalDeaths { get; set; }
-        public int TotalShotsHit { get; set; }
-        public int TotalShotsFired { get; set; }
+        public double TotalDeaths { get; set; }
+        public double TotalShotsHit { get; set; }
+        public double TotalShotsFired { get; set; }
         public double Accuracy { get; set; }
         public double HeadshotAccuracy { get; set; }
-        public int TotalTimePlayed { get; set; }
-        public int TotalMatchesWon { get; set; }
-        public int TotalMatchesPlayed { get; set; }
+        public double TotalTimePlayed { get; set; }
+        public double TotalMatchesWon { get; set; }
+        public double TotalMatchesPlayed { get; set; }
         public double MatchesWinPercentage { get; set; }
-        public int TotalGunGameMatchesWon { get; set; }
-        public int TotalGunGameMatchesPlayed { get; set; }
+        public double TotalGunGameMatchesWon { get; set; }
+        public double TotalGunGameMatchesPlayed { get; set; }
         public double GunGameMatchesWinPercentage { get; set; }
-        public int TotalWeaponsDonated { get; set; }
-        public int TotalBrokenWindows { get; set; }
-        public int TotalDominations { get; set; }
-        public int TotalDominationOverkills { get; set; }
-        public int TotalRevenges { get; set; }
-        public int TotalPlantedBombs { get; set; }
-        public int TotalDefusedBombs { get; set; }
-        public int TotalWins { get; set; }
-        public int TotalDamageDone { get; set; }
-        public int TotalMoneyEarned { get; set; }
-        public int TotalRescuedHostages { get; set; }
-        public int TotalRoundsPlayed { get; set; }
-        public int TotalMVPs { get; set; }
-        public int TotalContributionScore { get; set; }
-        public int TotalGunGameContributionScore { get; set; }
-        public int TotalGunGameRoundsWon { get; set; }
-        public int TotalGunGameRoundsPlayed { get; set; }
+        public double TotalWeaponsDonated { get; set; }
+        public double TotalBrokenWindows { get; set; }
+        public double TotalDominations { get; set; }
+        public double TotalDominationOverkills { get; set; }
+        public double TotalRevenges { get; set; }
+        public double TotalPlantedBombs { get; set; }
+        public double TotalDefusedBombs { get; set; }
+        public double TotalWins { get; set; }
+        public double TotalDamageDone { get; set; }
+        public double TotalMoneyEarned { get; set; }
+        public double TotalRescuedHostages { get; set; }
+        public double TotalRoundsPlayed { get; set; }
+        public double TotalMVPs { get; set; }
+        public double TotalContributionScore { get; set; }
+        public double TotalGunGameContributionScore { get; set; }
+        public double TotalGunGameRoundsWon { get; set; }
+        public double TotalGunGameRoundsPlayed { get; set; }
         // Todo: total_progressive_matches_won
         // Todo: total_TR_planted_bombs
         // Todo: total_TR_defused_bombs
